+++ nym-vpn-core/crates/nym-vpn-lib/src/service/config/config_manager.rs
@@ -10,9 +10,9 @@ use std::{
 use nym_common::trace_err_chain;
 use nym_http_api_client::{Client, FrontPolicy};
 use nym_registration_client::MixnetClientConfig;
+use nym_vpn_lib_types::MixnetTrafficConfigValidationError;
 #[cfg(any(target_os = "macos", target_os = "windows"))]
 use nym_vpn_lib_types::SplitApp;
-use nym_vpn_lib_types::{GatewayIndependence, MixnetTrafficConfigValidationError};
 use tokio::{fs, sync::broadcast};
 
 use crate::{
@@ -258,8 +258,19 @@ impl VpnServiceConfigManager {
         if (enable_gateway_independence && !self.config.gateway_independence.full_enabled())
             || (!enable_gateway_independence && !self.config.gateway_independence.full_disabled())
         {
-            self.config.gateway_independence =
-                GatewayIndependence::new(enable_gateway_independence);
+            self.config
+                .gateway_independence
+                .set_enabled(enable_gateway_independence);
+            self.save_config_and_send_event().await;
+        }
+    }
+
+    pub async fn set_enable_gateway_independence_notifications(
+        &mut self,
+        enable_notifications: bool,
+    ) {
+        if enable_notifications != self.config.gateway_independence.enable_notifications {
+            self.config.gateway_independence.enable_notifications = enable_notifications;
             self.save_config_and_send_event().await;
         }
     }